Feat/extract pitch module (#398)
* feat: extract pitch from core package

* feat: add pitch as dependency

* chore: bump version
