+++ packages/core/index.ts
@@ -1,6 +1,6 @@
-export * from "./src/utils";
-export * from "./src/named";
-export * from "./src/pitch";
-export * from "./src/note";
-export * from "./src/interval";
+export * from "@tonaljs/pitch";
 export * from "./src/distance";
+export * from "./src/interval";
+export * from "./src/named";
+export * from "./src/note";
+export * from "./src/utils";
