--- packages/core/src/pitch.ts
@@ -1,77 +1,0 @@
-type Fifths = number;
-type Octaves = number;
-export type Direction = 1 | -1;
-
-export type PitchClassCoordinates = [Fifths];
-export type NoteCoordinates = [Fifths, Octaves];
-export type IntervalCoordinates = [Fifths, Octaves, Direction];
-export type PitchCoordinates =
-  | PitchClassCoordinates
-  | NoteCoordinates
-  | IntervalCoordinates;
-
-/**
- * Pitch properties
- *
- * - {number} step - The step number: 0 = C, 1 = D, ... 6 = B
- * - {number} alt - Number of alterations: -2 = 'bb', -1 = 'b', 0 = '', 1 = '#', ...
- * - {number} [oct] = The octave (undefined when is a coord class)
- * - {number} [dir] = Interval direction (undefined when is not an interval)
- */
-export interface Pitch {
-  readonly step: number;
-  readonly alt: number;
-  readonly oct?: number; // undefined for pitch classes
-  readonly dir?: Direction; // undefined for notes
-}
-
-export function isPitch(pitch: unknown): pitch is Pitch {
-  return pitch !== null &&
-    typeof pitch === "object" &&
-    "step" in pitch &&
-    typeof pitch.step === "number" &&
-    "alt" in pitch &&
-    typeof pitch.alt === "number"
-    ? true
-    : false;
-}
-
-// The number of fifths of [C, D, E, F, G, A, B]
-const FIFTHS = [0, 2, 4, -1, 1, 3, 5];
-// The number of octaves it span each step
-const STEPS_TO_OCTS = FIFTHS.map((fifths: number) =>
-  Math.floor((fifths * 7) / 12),
-);
-
-export function encode(pitch: Pitch): PitchCoordinates {
-  const { step, alt, oct, dir = 1 } = pitch;
-  const f = FIFTHS[step] + 7 * alt;
-  if (oct === undefined) {
-    return [dir * f];
-  }
-  const o = oct - STEPS_TO_OCTS[step] - 4 * alt;
-  return [dir * f, dir * o];
-}
-
-// We need to get the steps from fifths
-// Fifths for CDEFGAB are [ 0, 2, 4, -1, 1, 3, 5 ]
-// We add 1 to fifths to avoid negative numbers, so:
-// for ["F", "C", "G", "D", "A", "E", "B"] we have:
-const FIFTHS_TO_STEPS = [3, 0, 4, 1, 5, 2, 6];
-
-export function decode(coord: PitchCoordinates): Pitch {
-  const [f, o, dir] = coord;
-  const step = FIFTHS_TO_STEPS[unaltered(f)];
-  const alt = Math.floor((f + 1) / 7);
-  if (o === undefined) {
-    return { step, alt, dir };
-  }
-  const oct = o + 4 * alt + STEPS_TO_OCTS[step];
-  return { step, alt, oct, dir };
-}
-
-// Return the number of fifths as if it were unaltered
-function unaltered(f: number): number {
-  const i = (f + 1) % 7;
-  return i < 0 ? 7 + i : i;
-}
