+++ packages/pitch/CHANGELOG.md
@@ -1,33 +1,5 @@
 # @tonaljs/pitch
 
-## 4.8.0
+## 5.0.0
 
-### Minor Changes
-
-- - fix time signature parsing
-  - add support for irrational time signatures
-  - add option `assumePerfectFifth` to `Chord.detect` function
-
-## 4.7.2
-
-### Patch Changes
-
-- Unify package versions
-
-## 4.6.5
-
-### Patch Changes
-
-- fix npm publish problem
-
-## 4.6.4
-
-### Patch Changes
-
-- fix memory leak
-
-## 4.6.3
-
-### Patch Changes
-
-- Bug fixing
+First release
