+++ packages/pitch/index.ts
@@ -1,29 +1,29 @@
-export type PitchClass = {
-  readonly step: number;
-  readonly alt: number;
-  readonly oct?: undefined;
-  readonly dir?: undefined;
-};
-export type NotePitch = {
-  readonly step: number;
-  readonly alt: number;
-  readonly oct: number;
-  readonly dir?: undefined;
-};
-export type IntervalPitch = {
-  readonly step: number;
-  readonly alt: number;
-  readonly oct: number;
-  readonly dir: 1 | -1;
-};
-export type Pitch = PitchClass | NotePitch | IntervalPitch;
+type Fifths = number;
+type Octaves = number;
+export type Direction = 1 | -1;
 
-export type InvalidPitch = {
-  readonly step?: undefined;
-  readonly alt?: undefined;
-  readonly oct?: undefined;
-  readonly dir?: undefined;
-};
+export type PitchClassCoordinates = [Fifths];
+export type NoteCoordinates = [Fifths, Octaves];
+export type IntervalCoordinates = [Fifths, Octaves, Direction];
+export type PitchCoordinates =
+  | PitchClassCoordinates
+  | NoteCoordinates
+  | IntervalCoordinates;
+
+/**
+ * Pitch properties
+ *
+ * - {number} step - The step number: 0 = C, 1 = D, ... 6 = B
+ * - {number} alt - Number of alterations: -2 = 'bb', -1 = 'b', 0 = '', 1 = '#', ...
+ * - {number} [oct] = The octave (undefined when is a coord class)
+ * - {number} [dir] = Interval direction (undefined when is not an interval)
+ */
+export interface Pitch {
+  readonly step: number;
+  readonly alt: number;
+  readonly oct?: number; // undefined for pitch classes
+  readonly dir?: Direction; // undefined for notes
+}
 
 const SIZES = [0, 2, 4, 5, 7, 9, 11];
 export const chroma = ({ step, alt }: Pitch) => (SIZES[step] + alt + 120) % 12;
@@ -36,11 +36,16 @@ export const midi = (pitch: Pitch) => {
   return pitch.oct !== undefined && h >= -12 && h <= 115 ? h + 12 : null;
 };
 
-export const freq = (pitch: Pitch) =>
-  pitch.oct !== undefined ? Math.pow(2, (height(pitch) - 57) / 12) * 440 : null;
-
-// Coordinates are [fifths] (for pitch classes) and [fifths, octaves] for notes and intervals
-export type PitchCoordinates = [number] | [number, number];
+export function isPitch(pitch: unknown): pitch is Pitch {
+  return pitch !== null &&
+    typeof pitch === "object" &&
+    "step" in pitch &&
+    typeof pitch.step === "number" &&
+    "alt" in pitch &&
+    typeof pitch.alt === "number"
+    ? true
+    : false;
+}
 
 // The number of fifths of [C, D, E, F, G, A, B]
 const FIFTHS = [0, 2, 4, -1, 1, 3, 5];
@@ -49,6 +54,9 @@ const STEPS_TO_OCTS = FIFTHS.map((fifths: number) =>
   Math.floor((fifths * 7) / 12),
 );
 
+/**
+ * Get coordinates from pitch object
+ */
 export function coordinates(pitch: Pitch): PitchCoordinates {
   const { step, alt, oct, dir = 1 } = pitch;
   const f = FIFTHS[step] + 7 * alt;
@@ -65,15 +73,17 @@ export function coordinates(pitch: Pitch): PitchCoordinates {
 // for ["F", "C", "G", "D", "A", "E", "B"] we have:
 const FIFTHS_TO_STEPS = [3, 0, 4, 1, 5, 2, 6];
 
+/**
+ * Get pitch from coordinate objects
+ */
 export function pitch(coord: PitchCoordinates): Pitch {
-  const [f, o] = coord;
+  const [f, o, dir] = coord;
   const step = FIFTHS_TO_STEPS[unaltered(f)];
   const alt = Math.floor((f + 1) / 7);
   if (o === undefined) {
-    return { step, alt };
+    return { step, alt, dir };
   }
   const oct = o + 4 * alt + STEPS_TO_OCTS[step];
-  const dir = undefined;
   return { step, alt, oct, dir };
 }
