+++ packages/pitch/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@tonaljs/pitch",
-  "private": true,
-  "version": "4.8.0",
+  "version": "5.0.0",
   "description": "Pitch in (western) music theory",
   "keywords": [
     "pitch",
