+++ packages/pitch/test.ts
@@ -1,11 +1,4 @@
-import {
-  IntervalPitch,
-  chroma,
-  coordinates,
-  height,
-  midi,
-  pitch,
-} from "./index";
+import { Pitch, chroma, coordinates, height, midi, pitch } from "./index";
 
 // Pitch classes
 const C = { step: 0, alt: 0 };
@@ -19,8 +12,8 @@ const A4 = { step: 5, alt: 0, oct: 4 };
 const Gs6 = { step: 4, alt: 1, oct: 6 };
 
 // Intervals
-const P5: IntervalPitch = { step: 4, alt: 0, oct: 0, dir: 1 };
-const P_5: IntervalPitch = { step: 4, alt: 0, oct: 0, dir: -1 };
+const P5: Pitch = { step: 4, alt: 0, oct: 0, dir: 1 };
+const P_5: Pitch = { step: 4, alt: 0, oct: 0, dir: -1 };
 
 describe("@tonaljs/pitch", () => {
   test("height", () => {
