+++ packages/tonal/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tonal",
-  "version": "5.1.0",
+  "version": "5.1.1",
   "description": "tonaljs music theory library",
   "keywords": [
     "music",
@@ -21,7 +21,7 @@
     "@tonaljs/chord-type": "^5.0.2",
     "@tonaljs/chord": "^5.0.0",
     "@tonaljs/collection": "^4.8.0",
-    "@tonaljs/core": "^4.10.0",
+    "@tonaljs/core": "^4.10.1",
     "@tonaljs/duration-value": "^4.8.0",
     "@tonaljs/interval": "^4.8.0",
     "@tonaljs/key": "^4.9.1",
